Added first res dialog, doesn't work in the Chasm, maybe fixed by build
The dialogue functions correctly when injected in the trigger on the tutorial level, but doesn't seem to work in the Chasm. 
Suspected cause : the absence of initialisation of the dialogue subsystems in the level blueprint of the Chasm, while the dialogue subsystems are mentioned in the tutorial level blueprint.
Does the issue disappear when the game is built, because Aaron starts in the Tutorial level?